--- a/fabio.properties
+++ b/fabio.properties
@@ -302,6 +302,20 @@
 # proxy.header.tls.value =
 
 
+# proxy access log
+# Example:
+#     proxy.log.enable = true
+#     proxy.log.file.path = access.log
+#     proxy.log.format = $remote_addr $time $request $body_bytes_sent $http_referer $http_user_agent $server_name
+#     $upstream_addr $upstream_response_time $request_args
+#
+# The default is
+#
+# proxy.log.enable =  false
+# proxy.log.file.path =
+# proxy.log.format =
+
+
 # registry.backend configures which backend is used.
 # Supported backends are: consul, static, file
 #