--- a/fabio.properties
+++ b/fabio.properties
@@ -306,8 +306,7 @@
 # Example:
 #     proxy.log.enable = true
 #     proxy.log.file.path = access.log
-#     proxy.log.format = $remote_addr $time $request $body_bytes_sent $http_referer $http_user_agent $server_name
-#     $upstream_addr $upstream_response_time $request_args
+#     proxy.log.format = $remote_addr $time $request $body_bytes_sent $http_referer $http_user_agent $server_name $upstream_addr $upstream_response_time $request_args
 #
 # The default is
 #