--- a/fabio.properties
+++ b/fabio.properties
@@ -304,14 +304,12 @@
 
 # proxy access log
 # Example:
-#     proxy.log.enable = true
-#     proxy.log.file.path = access.log
-#     proxy.log.format = $remote_addr $time $request $body_bytes_sent $http_referer $http_user_agent $server_name $upstream_addr $upstream_response_time $request_args
+#     proxy.log.target = stdout 
+#     proxy.log.format = remote_addr time request body_bytes_sent http_referer http_user_agent server_name proxy_endpoint response_time request_args
 #
 # The default is
 #
-# proxy.log.enable =  false
-# proxy.log.file.path =
+# proxy.log.target =
 # proxy.log.format =