--- a/fabio.properties
+++ b/fabio.properties
@@ -302,10 +302,11 @@
 # proxy.header.tls.value =
 
 
-# proxy access log
+# proxy log configures which access logger is used.
+#
 # Example:
 #     proxy.log.target = stdout 
-#     proxy.log.format = remote_addr time request body_bytes_sent http_referer http_user_agent server_name proxy_endpoint response_time request_args
+#     proxy.log.format = remote_addr time request body_bytes_sent http_referer http_user_agent server_name proxy_endpoint response_time request_args http_x_forwarded_for
 #
 # The default is
 #