--- a/proxy/logger.go
+++ b/proxy/logger.go
@@ -1,147 +1,113 @@
 package proxy
 
 import (
+	"bytes"
 	"fmt"
+	"io"
 	"log"
 	"net/http"
+	"os"
 	"strings"
+	"sync"
 	"time"
+
+	"github.com/eBay/fabio/config"
 )
 
-var (
-	pattern *Pattern
-)
-
-type Pattern struct {
-	converts []Converter
+type Logger struct {
+	mu      sync.Mutex // ensures atomic writes;
+	out     io.Writer  // destination for output
+	pattern []Pattern
+	enable  bool
 }
 
-type Converter interface {
-	convert(t time.Time, r http.Request) string
-}
+func newLogger(cfg config.Proxy) *Logger {
 
-type RequestConvert struct {
-}
+	if len(cfg.Log.Target) > 0 && len(cfg.Log.Format) <= 0 {
+		log.Fatal("[FATAL] Invalid Logger format ", cfg.Log.Format)
+	}
 
-func (RequestConvert) convert(t time.Time, r http.Request) string {
-	return "\"" + r.Method + " " + r.URL.Path + " " + r.Proto + "\""
-}
-
-type RemoteAddrConvert struct {
-}
-
-func (RemoteAddrConvert) convert(t time.Time, r http.Request) string {
-	return r.RemoteAddr[:strings.Index(r.RemoteAddr, ":")]
-}
-
-type BodyBytesSentConvert struct {
-}
-
-func (BodyBytesSentConvert) convert(t time.Time, r http.Request) string {
-	return fmt.Sprintf("%d", uint64(r.ContentLength))
-}
-
-type HttpRefererConvert struct {
-}
-
-func (HttpRefererConvert) convert(t time.Time, r http.Request) string {
-	return r.Referer()
-}
-
-type HttpUserAgentConvert struct {
-}
-
-func (HttpUserAgentConvert) convert(t time.Time, r http.Request) string {
-	return r.UserAgent()
-}
-
-type ServerNameConvert struct {
-}
-
-func (ServerNameConvert) convert(t time.Time, r http.Request) string {
-	return r.Host
-}
-
-type UpstreamAddrConvert struct {
-}
-
-func (UpstreamAddrConvert) convert(t time.Time, r http.Request) string {
-	return r.URL.Host
-}
-
-type RequestArgsConvert struct {
-}
-
-func (RequestArgsConvert) convert(t time.Time, r http.Request) string {
-	return r.URL.RawQuery
-}
-
-type UpstreamResponseTimeConvert struct {
-}
-
-func (UpstreamResponseTimeConvert) convert(t time.Time, r http.Request) string {
-	return fmt.Sprintf("%.4f", time.Since(t).Seconds())
-}
-
-type TimeConvert struct {
-}
-
-func (TimeConvert) convert(t time.Time, r http.Request) string {
-	return t.Format(time.RFC3339)
-}
-
-type DefaultConvert struct {
-}
-
-func (DefaultConvert) convert(t time.Time, r http.Request) string {
-	return ""
-}
-
-func patternParse(pattern string) *Pattern {
-	flags := strings.Split(pattern, " ")
-
-	c := make([]Converter, len(flags))
+	flags := strings.Split(cfg.Log.Format, " ")
+	c := make([]Pattern, len(flags))
 
 	for i, flag := range flags {
-		switch flag {
-		case "$remote_addr":
-			c[i] = &RemoteAddrConvert{}
-		case "$time":
-			c[i] = &TimeConvert{}
-		case "$request":
-			c[i] = &RequestConvert{}
-		case "$body_bytes_sent":
-			c[i] = &BodyBytesSentConvert{}
-		case "$http_referer":
-			c[i] = &HttpRefererConvert{}
-		case "$http_user_agent":
-			c[i] = &HttpUserAgentConvert{}
-		case "$server_name":
-			c[i] = &ServerNameConvert{}
-		case "$upstream_addr":
-			c[i] = &UpstreamAddrConvert{}
-		case "$upstream_response_time":
-			c[i] = &UpstreamResponseTimeConvert{}
-		case "$request_args":
-			c[i] = &RequestArgsConvert{}
-		default:
-			c[i] = &DefaultConvert{}
-		}
+		c[i] = patterns[flag]
 	}
 
-	return &Pattern{converts: c}
+	switch cfg.Log.Target {
+	case "stdout":
+		log.Printf("[INFO] Output logger to stdout")
+		return &Logger{out: os.Stdout, pattern: c, enable: true}
+	case "":
+		log.Printf("[INFO] Logger disabled")
+	default:
+		log.Fatal("[FATAL] Invalid Logger target ", cfg.Log.Target)
+	}
+	return &Logger{}
 }
 
-func logger(log *log.Logger, format string, t time.Time, r *http.Request) {
+type Pattern func(w io.Writer, t time.Time, r *http.Request)
 
-	if pattern == nil {
-		pattern = patternParse(format)
+const BufferSize = 1024
+
+var (
+	pool = sync.Pool{
+		New: func() interface{} {
+			return bytes.NewBuffer(make([]byte, 0, BufferSize))
+		},
+	}
+	patterns = map[string]Pattern{
+		"remote_addr": func(w io.Writer, t time.Time, r *http.Request) {
+			io.WriteString(w, r.RemoteAddr[:strings.Index(r.RemoteAddr, ":")])
+		},
+		"time": func(w io.Writer, t time.Time, r *http.Request) {
+			io.WriteString(w, t.Format(time.RFC3339))
+		},
+		"request": func(w io.Writer, t time.Time, r *http.Request) {
+			io.WriteString(w, "\""+r.Method+" "+r.URL.Path+" "+r.Proto+"\"")
+		},
+		"body_bytes_sent": func(w io.Writer, t time.Time, r *http.Request) {
+			io.WriteString(w, fmt.Sprintf("%d", uint64(r.ContentLength)))
+		},
+		"http_referer": func(w io.Writer, t time.Time, r *http.Request) {
+			io.WriteString(w, r.Referer())
+		},
+		"http_user_agent": func(w io.Writer, t time.Time, r *http.Request) {
+			io.WriteString(w, r.UserAgent())
+		},
+		"server_name": func(w io.Writer, t time.Time, r *http.Request) {
+			io.WriteString(w, r.Host)
+		},
+		"proxy_endpoint": func(w io.Writer, t time.Time, r *http.Request) {
+			io.WriteString(w, r.URL.Host)
+		},
+		"response_time": func(w io.Writer, t time.Time, r *http.Request) {
+			io.WriteString(w, fmt.Sprintf("%.4f", time.Since(t).Seconds()))
+		},
+		"request_args": func(w io.Writer, t time.Time, r *http.Request) {
+			io.WriteString(w, r.URL.RawQuery)
+		},
+	}
+)
+
+func (l *Logger) logger(t time.Time, r *http.Request) {
+
+	if !l.enable {
+		return
 	}
 
-	str := ""
-	for _, c := range pattern.converts {
-		str += c.convert(t, *r) + " "
-	}
+	l.mu.Lock()
+	defer l.mu.Unlock()
 
-	log.Println(str)
+	b := pool.Get().(*bytes.Buffer)
+	b.Reset()
+
+	for _, p := range l.pattern {
+		p(b, t, r)
+		b.WriteRune(' ')
+	}
+	b.Truncate(b.Len() - 1) //drop last space
+	b.WriteByte('\n')
+	l.out.Write(b.Bytes())
+	pool.Put(b)
 }