--- a/proxy/proxy.go
+++ b/proxy/proxy.go
@@ -1,9 +1,7 @@
 package proxy
 
 import (
-	"log"
 	"net/http"
-	"os"
 	"time"
 
 	"github.com/eBay/fabio/config"
@@ -16,24 +14,16 @@ type Proxy struct {
 	tr       http.RoundTripper
 	cfg      config.Proxy
 	requests gometrics.Timer
-	log      *log.Logger
+	log      *Logger
 }
 
 func New(tr http.RoundTripper, cfg config.Proxy) *Proxy {
-	var l *log.Logger
-	if cfg.Log.Enable {
-		file, err := os.OpenFile(cfg.Log.File.Path, os.O_CREATE|os.O_RDWR|os.O_APPEND, 0666)
-		if err != nil {
-			log.Fatal("[FATAL] ", err)
-		}
-		l = log.New(file, "", 0)
-	}
 
 	return &Proxy{
 		tr:       tr,
 		cfg:      cfg,
 		requests: gometrics.GetOrRegisterTimer("requests", gometrics.DefaultRegistry),
-		log:      l,
+		log:      newLogger(cfg),
 	}
 }
 
@@ -69,8 +59,5 @@ func (p *Proxy) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 	h.ServeHTTP(w, r)
 	p.requests.UpdateSince(start)
 	t.Timer.UpdateSince(start)
-
-	if p.cfg.Log.Enable {
-		logger(p.log, p.cfg.Log.Format, start, r)
-	}
+	p.log.logger(start, r)
 }