--- a/proxy/proxy.go
+++ b/proxy/proxy.go
@@ -1,10 +1,12 @@
 package proxy
 
 import (
+	"log"
 	"net/http"
 	"time"
 
 	"github.com/eBay/fabio/config"
+	"github.com/eBay/fabio/logger"
 
 	gometrics "github.com/rcrowley/go-metrics"
 )
@@ -14,16 +16,22 @@ type Proxy struct {
 	tr       http.RoundTripper
 	cfg      config.Proxy
 	requests gometrics.Timer
-	log      *Logger
+	logger   *logger.Logger
 }
 
 func New(tr http.RoundTripper, cfg config.Proxy) *Proxy {
+
+	logger, err := newLogger(cfg.Log.Target, cfg.Log.Format)
+
+	if err != nil {
+		log.Fatal("[FATAL] ", err)
+	}
 
 	return &Proxy{
 		tr:       tr,
 		cfg:      cfg,
 		requests: gometrics.GetOrRegisterTimer("requests", gometrics.DefaultRegistry),
-		log:      newLogger(cfg),
+		logger:   logger,
 	}
 }
 
@@ -59,5 +67,7 @@ func (p *Proxy) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 	h.ServeHTTP(w, r)
 	p.requests.UpdateSince(start)
 	t.Timer.UpdateSince(start)
-	p.log.logger(start, r)
+	if p.logger != nil {
+		p.logger.Log(start, r)
+	}
 }