--- a/proxy/proxy_integration_test.go
+++ b/proxy/proxy_integration_test.go
@@ -55,3 +55,33 @@ func TestProxyNoRouteStaus(t *testing.T) {
 		t.Fatalf("got %d want %d", got, want)
 	}
 }
+
+func BenchmarkProxyLogger(b *testing.B) {
+
+	got := "not called"
+	server := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		got = r.Header.Get("X-Forwarded-For")
+	}))
+	defer server.Close()
+
+	table := make(route.Table)
+	table.AddRoute("mock", "/", server.URL, 1, nil)
+	route.SetTable(table)
+
+	tr := &http.Transport{Dial: (&net.Dialer{}).Dial}
+	proxy := New(tr, config.Proxy{LocalIP: "1.1.1.1", ClientIPHeader: "X-Forwarded-For"})
+
+	req := &http.Request{
+		RequestURI: "/",
+		Header:     http.Header{"X-Forwarded-For": {"3.3.3.3"}},
+		RemoteAddr: "2.2.2.2:666",
+		URL:        &url.URL{},
+		Method:     "GET",
+		Proto:      "HTTP/1.1",
+	}
+
+	for i := 0; i < b.N; i++ {
+		proxy.ServeHTTP(httptest.NewRecorder(), req)
+	}
+
+}