Add access logging like #80 

Written to disk, format is configurable like nginx log_format

E.g
```
fabio.properties

proxy.log.enable = true
proxy.log.file.path = access.log
proxy.log.format = $remote_addr $time $request $body_bytes_sent $http_referer $http_user_agent $server_name $upstream_addr $upstream_response_time $request_args
```

support format syntax:

```
$remote_addr      
$time 
$request
$body_bytes_sent
$http_referer 
$http_user_agent 
$server_name 
$upstream_addr 
$upstream_response_time 
$request_args
```

But the response state does not support, because I do not know how to get.