@magiconair 2016-07-24
@beyondblog I had a look and here are my comments:

I like the approach of creating a list of functions which create the log line depending on the format. The implementation is abusing the types a bit and makes too many allocations by using string concatenation instead of a `bytes.Buffer` from a `sync.Pool`

I'd replace the types with something like this:

```
type value func(w io.Writer, t time.Time, r *http.Request)

var values = map[string]value{
	"$remote_addr": func(w io.Writer, t time.Time, r *http.Request) {
		w.WriteString(r.RemoteAddr[:strings.Index(r.RemoteAddr, ":")])
	},
	"$time": func(w io.Writer, t time.Time, r *http.Request) {
		w.WriteString(t.Format(time.RFC3339))
	},
	"$request": func(w io.Writer, t time.Time, r *http.Request) {
		w.WriteRune('\"')
		w.WriteString(r.Method)
		w.WriteRune(' ')
		w.WriteString(r.URL.Path)
		w.WriteRune(' ')
		w.Write(r.Proto)
		w.WriteRune('\"')
	},
}

var pattern []value
```

In addition, I'd use a `sync.Pool` to re-use the buffers since this code sits in the hot-path. 

```
var BufSize = 1024

var pool = sync.Pool {
	New:func() interface{} {
		return bytes.NewBuffer(make([]byte, 0, BufSize))
	},
}

func logger(w io.Writer, format string, t time.Time, r *http.Request) {
	b := pool.Get().(*bytes.Buffer)
	b.Reset()
	for _, p := range pattern {
		p(b, t, r)
		b.WriteRune(' ')
	}
	b.Truncate(b.Len()-1) // drop last space
	_, err := w.Write(b.Bytes())
}
```

I'd also drop the lazy pattern parsing and prefer an `Init()` function which parses the pattern and returns an error. This way you save a `nil` check on every request.

Then, I would not use a `log.Logger` for the output since you are already creating the full log line including time stamp. Just pass in an `io.Writer`.

I am also not sure if fabio should write the access log to a file. The normal log is written to `stderr` which means you need to capture and redirect it anyway. `fabio` should also be run under a supervisor which restarts it (daemontools, systemd, supervisor, ...) so that you could redirect the output to a file and let someone else handle log rotation. What if you want to ship the access logs via rsyslog over the network? What if fabio runs in a docker container. For now I would just send them to `stdout` which would save another allocation since you can then write the content of the buffer directly to stdout instead of converting it to a string first.

`proxy.log.enable` is redundant. `proxy.log.file.path` should be named `proxy.log.target` and should default to empty which disables access logging. The other option should be `stdout`. 

I might also drop the `$` from the format variables since it doesn't serve any purpose but it might increase readability. 

I might want to have a look at the speed of `time.Format(time.RFC3339)` and check if a special implementation would be better. Again, keep in mind that we're driving fabio with a couple of thousand requests per second and adding lots of allocations to every request will just put more pressure on the GC.