@magiconair 2016-07-25
Looks better. I've added some comments. Please drop the `enabled` flag since it is redundant. If you don't want to log you should not have a logger in the first place. I'd also move the logger creation into the `main` function since it can `FATAL` and then inject it into the Proxy.