@magiconair 2016-07-26
@beyondblog I think a logger package is a good idea. Maybe like this?

```
package logger

var BufSize = 1024

type value func(b *bytes.Buffer, t time.Time, r *http.Request)

var values = map[string]value { ... }
 
type Logger struct {
    // p stores the log pattern as a list of value functions
    p []value

    // mu guards w
    mu sync.Mutex

    // w is the log destination
    w io.Writer
}

func (l *Logger) New(w io.Writer, format string) (*Logger, error) {
    if w == nil { return nil, nil }
    p, err := parse(format)
    if err != nil { return nil, err }
    return &Logger{p, w}
}

func parse(format string) ([]value, error) {
    var vv []value
    for _, f := range strings.Fields(format) {
        v := values[f]
        if v == nil { return nil, fmt.Errorf("invalid log field %s", f) }
        vv = append(vv, v)
    }
    return vv, nil
}

func (l *Logger) Log(t time.Time, r *http.Request) {
    b := pool.Get()
    b.Reset()
    for _, v := range l.p {
        v(b, t, r)
        b.WriteRune(' ')
    }
    b.Truncate(b.Len()-1) // drop last space
    b.WriteRune('\n')
    l.mu.Lock()
    l.w.Write(b.Bytes())
    l.mu.Unlock()
    pool.Put(b)
}
```

Could you also add some table driven tests for the value functions and the parsing? I usually write them like this:

```
func TestLog(t *testing.T) {
	ts := time.Date(2016, 1, 1, 0, 0, 0, 0, time.UTC)
	tests := []struct {
		req    *http.Request
		format string
		out    string
	}{
		{},
		{&http.Request{Host: "abc.com"}, "req_host", "abc.com"},
                ...
	}

	for i, tt := range tests {
                b := new(bytes.Buffer)
		l, err := logger.New(b, tt.format)
                if err != nil { t.Fatalf("%d: got %v want nil", i, err) }
                l.Log(ts, tt.req)
		if got, want := string(b.Bytes()), tt.out; got != want {
			t.Errorf("%d: got %q want %q", i, got, want)
		}
	}
}
```