@magiconair 2016-07-26
Apologies to the watchers for spamming you with code-review comments. I'm used to the gerrit workflow where all comments get published in one go.