@beyondblog 2016-07-26
@magiconair Thanks. I will improve my code under your suggestion.