@magiconair 2016-08-01
@beyondblog I've taken your PR and refactored it a bit since the logger is sitting in the hot path and I'd prefer it not to allocate memory since that would put pressure on the GC. 

Unfortunately, I still haven't figured out how to collaborate on PRs the same way as it is possible to do this on gerrit. Therefore, I've created a new branch, squashed your commits and rebased them on `master` and added my cleanups on top of them. If you know how to do this better on GH then pls let me know.

Here is what I did:

* Add a benchmark and added some comparisons
* Use `bytes.Buffer` instead of `io.Writer` so that I can use the `WriteXXX` methods directly
* Inlined all calls to `fmt.Sprintf()` and `time.Format()` to avoid allocations. For this I've written a custom `atoi` function which does not alloc. 
* I've added more tests to cover all log fields and I am not sure if the way this is currently implemented yields to the correct results. However, I was focussing mainly on having the tests and the benchmarks and not on the correctness of the tests.

```
# before
BenchmarkLog-8	 1000000	      1807 ns/op	      80 B/op	       6 allocs/op

#after
BenchmarkLog-8	 2000000	       919 ns/op	       0 B/op	       0 allocs/op

$ benchcmp 0-baseline.txt 3.5-inline-time-format.txt
benchmark          old ns/op     new ns/op     delta
BenchmarkLog-8     1807          919           -49.14%

benchmark          old allocs     new allocs     delta
BenchmarkLog-8     6              0              -100.00%

benchmark          old bytes     new bytes     delta
BenchmarkLog-8     80            0             -100.00%
```