@beyondblog 2016-08-03
@magiconair  Thanks. I look forward to this feature.