@roman-gliffy 2016-10-15
Would be nice to have logs.
@magiconair, any suggestions how to deal with the absent of logs in production for the time being?